+++ src/main/java/me/anedhel/lotr/item/ModGearType.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright (c)
+ * Authors/Developers are listed in the CONTRIBUTING.md
+ *
+ * lord-of-the-rings-middle-earth is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * lord-of-the-rings-middle-earth is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * You should have received a copy of the GNU General Public License
+ *  along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package me.anedhel.lotr.item;
 
 import net.minecraft.item.Item;
@@ -7,9 +24,14 @@ import org.jetbrains.annotations.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Enum representing the different types of gear in the mod. A gear type is defined by its material, armor pieces, and
+ * tools.
+ */
 public enum ModGearType {
     BRONZE(ModItems.BRONZE_INGOT, ModItems.BRONZE_HELMET, ModItems.BRONZE_CHESTPLATE, ModItems.BRONZE_LEGGINGS, ModItems.BRONZE_BOOTS,
-            ModItems.BRONZE_SWORD, ModItems.BRONZE_AXE, ModItems.BRONZE_PICKAXE, ModItems.BRONZE_SHOVEL, ModItems.BRONZE_HOE, false);
+            ModItems.BRONZE_SWORD, ModItems.BRONZE_AXE, ModItems.BRONZE_PICKAXE, ModItems.BRONZE_SHOVEL,
+            ModItems.BRONZE_HOE);
 
     private final Item material;
     private final Item helmet;
@@ -21,9 +43,9 @@ public enum ModGearType {
     private final Item pickaxe;
     private final Item shovel;
     private final Item hoe;
-    private final boolean vanillaAddition;
 
-    ModGearType(Item material, Item helmet, Item chestplate, Item leggings, Item boots, Item sword, Item axe, Item pickaxe, Item shovel, Item hoe, boolean vanillaAddition) {
+    ModGearType(Item material, Item helmet, Item chestplate, Item leggings, Item boots, Item sword, Item axe,
+            Item pickaxe, Item shovel, Item hoe) {
         this.material = material;
         this.helmet = helmet;
         this.chestplate = chestplate;
@@ -34,9 +56,13 @@ public enum ModGearType {
         this.pickaxe = pickaxe;
         this.shovel = shovel;
         this.hoe = hoe;
-        this.vanillaAddition = vanillaAddition;
     }
 
+    /**
+     * Returns a list of all items in the tool item group.
+     *
+     * @return a list of all items in the tool item group
+     */
     public @NotNull List<ItemStack> getToolItemGroupList () {
         List<ItemStack> toolItemGroupList = new ArrayList<>();
         if (axe != null) {
@@ -54,6 +80,10 @@ public enum ModGearType {
         return toolItemGroupList;
     }
 
+    /**
+     * Returns a list of all items in the combat item group.
+     * @return a list of all items in the combat item group
+     */
     public @NotNull List<ItemStack> getCombatItemGroupList () {
         List<ItemStack> combatItemGroupList = new ArrayList<>();
         if (sword != null) {
@@ -112,9 +142,5 @@ public enum ModGearType {
 
     public Item getHoe() {
         return hoe;
-    }
-
-    public boolean isVanillaAddition() {
-        return vanillaAddition;
     }
 }